+++ src/main/java/com/feibijiubi/backend/service/auth/TokenService.java
@@ -5,5 +5,5 @@ import com.feibijiubi.backend.entity.User;
 public interface TokenService {
     String createToken(User user);
 
-    Long getUserId(String token);
+    Integer getUserId(String token);
 }
